refector: video & ImageGallery Type에 맞게 컴포넌트 불러오는 형태로 수정
