+++ online_invitaion/src/components/sections/contacts/Contact.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames/bind'
 import Section from '@/components/shared/sections/Section'
 import styles from './Contact.module.scss'
 import Accordion from '../../shared/accordions/Accordion'
-import { Person, Wedding } from '@models/wedding'
+import { Person, Wedding } from '@/models/invitation'
 import { CopyToClipboard } from 'react-copy-to-clipboard'
 
 const cx = classNames.bind(styles)
