+++ online_invitaion/src/components/sections/videos/Type001/index.tsx
@@ -1,8 +1,9 @@
+import React from 'react'
 import classNames from 'classnames/bind'
 import styles from './Video.module.scss'
 
 // 공통 섹션 디자인 불러오기
-import Section from '@/components/shared/sections/Section'
+import Section from '@components/shared/sections/Section'
 
 const cx = classNames.bind(styles)
